no se

+++ cub3d.h
@@ -70,8 +70,8 @@ typedef struct s_raycasting
 	double	delta_dist_x;
 	double	delta_dist_y;
 	double	perp_wall_dist;
-	int		step_x;
-	int		step_y;
+	size_t step_x;
+	size_t step_y;
 	int		hit;
 	int		side;
 	int		line_height;
